+++ sdk/network/Azure.ResourceManager.Network/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- For the `HopLink` class, Use the new property `ResourceIdString` to replace `ResourceId` and fix the formatting issue when `resourceId` returns an empty string.
+
 ### Bugs Fixed
 
 ### Other Changes